requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
kombu==4.6.8
pytz==2020.1
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
